Move configs to messaging settings
