+++ streampipes-model/src/main/java/org/apache/streampipes/model/configuration/SpCoreConfiguration.java
@@ -32,23 +32,6 @@ public class SpCoreConfiguration {
   private EmailConfig emailConfig;
   private GeneralConfig generalConfig;
 
-  private String jmsHost;
-  private int jmsPort;
-
-  private String mqttHost;
-  private int mqttPort;
-
-  private String natsHost;
-  private int natsPort;
-
-  private String kafkaHost;
-  private int kafkaPort;
-
-  private String pulsarUrl;
-
-  private String zookeeperHost;
-  private int zookeeperPort;
-
   private boolean isConfigured;
 
   private String assetDir;
@@ -103,94 +86,6 @@ public class SpCoreConfiguration {
 
   public void setGeneralConfig(GeneralConfig generalConfig) {
     this.generalConfig = generalConfig;
-  }
-
-  public String getJmsHost() {
-    return jmsHost;
-  }
-
-  public void setJmsHost(String jmsHost) {
-    this.jmsHost = jmsHost;
-  }
-
-  public int getJmsPort() {
-    return jmsPort;
-  }
-
-  public void setJmsPort(int jmsPort) {
-    this.jmsPort = jmsPort;
-  }
-
-  public String getMqttHost() {
-    return mqttHost;
-  }
-
-  public void setMqttHost(String mqttHost) {
-    this.mqttHost = mqttHost;
-  }
-
-  public int getMqttPort() {
-    return mqttPort;
-  }
-
-  public void setMqttPort(int mqttPort) {
-    this.mqttPort = mqttPort;
-  }
-
-  public String getNatsHost() {
-    return natsHost;
-  }
-
-  public void setNatsHost(String natsHost) {
-    this.natsHost = natsHost;
-  }
-
-  public int getNatsPort() {
-    return natsPort;
-  }
-
-  public void setNatsPort(int natsPort) {
-    this.natsPort = natsPort;
-  }
-
-  public String getKafkaHost() {
-    return kafkaHost;
-  }
-
-  public void setKafkaHost(String kafkaHost) {
-    this.kafkaHost = kafkaHost;
-  }
-
-  public int getKafkaPort() {
-    return kafkaPort;
-  }
-
-  public void setKafkaPort(int kafkaPort) {
-    this.kafkaPort = kafkaPort;
-  }
-
-  public String getPulsarUrl() {
-    return pulsarUrl;
-  }
-
-  public void setPulsarUrl(String pulsarUrl) {
-    this.pulsarUrl = pulsarUrl;
-  }
-
-  public String getZookeeperHost() {
-    return zookeeperHost;
-  }
-
-  public void setZookeeperHost(String zookeeperHost) {
-    this.zookeeperHost = zookeeperHost;
-  }
-
-  public int getZookeeperPort() {
-    return zookeeperPort;
-  }
-
-  public void setZookeeperPort(int zookeeperPort) {
-    this.zookeeperPort = zookeeperPort;
   }
 
   public boolean isConfigured() {
